Abstract questionnaire mapping methods to external module
